move waiting list back to administration
