--- app/views/registrations/waiting_list.html.erb
@@ -1,5 +1,0 @@
-<% provide(:title, I18n.t('registrations.list.waiting_list', comp: @competition.name)) %>
-
-<%= render layout: "nav" do %>
-  <%= react_component('RegistrationsV2/Waiting', { competitionInfo: @competition.to_competition_info }) %>
-<% end %>
